Minor changes to improve output formatting.
Modified command name and added listing of each model's methods


Corrected order for entry of command in the docs


Removed extra spaces on line endings.
